requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
openai>=1.50.0
anthropic>=0.36.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
pypdf>=5.0.0
unstructured>=0.16.0
httpx>=0.27.0
rich>=13.9.0
typer>=0.13.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
